砲撃エフェクトの仮実装。アニメーションの描画は一応実装完了。 索敵エフェクトを作っていたら、ソナー専用エフェクトのようなものが完成してしまった。

+++ BattleShip/BattleShip/src/GameLog.cpp
@@ -20,7 +20,7 @@ void GameLog::AddStream( const char* _str, const D3DXCOLOR& _color, const unsign
 		tempY += logBack->GetHeight();	///< Šî–{‚‚³‚µ‚©•Ï‚í‚ç‚È‚¢‚Ì‚ÅYÀ•W‚¾‚¯•¶Žš‚Ìc••ª‰ÁŽZ
 	}
 	m_logStream.push_back( new LogStream( _str, tempX, tempY, _width, _height, _color ) );
-	CheckStream( _LOG_HEIGHT_MAX_ );
+	CheckStream( static_cast<unsigned int>(_LOG_HEIGHT_MAX_) );
 }
